Populate plate metadata collection directly DbResultWriter
